DailyView fix
modified how the axios errors are handled per the results of the unit
tests. When the error was being caught, it was still passing on to the
subsequent .then() calls (which is how promises work). As such, I had
to remove the inner errorHandler functions into a catch() call at the
end of the chain of then()s.