DailyView integration tests
Looking at the coverage metrics for my modules seems to highlight some
of the limitations of coverage metrics. In particular, I have achieved
good numbers for my EntryCreator widget, in spite of not writing a
single unit test for it. I have also retained poor % Branch metrics for
the DailyView widget. While this is true, the branches are all just
like ‘else (default error message)’ which I suppose I *should* test,
but don’t feel like writing all that code for. I have also noticed that
coverage does not indicate props passing correctness or things of that
ilk, which makes sense, but is just something to keep in mind.